webcrawler/spiders/requirements.txt
Scrapy
py-term
